fix(frankenlibc-abi): tighten glibc ABI parity for scalb, math & string wrappers
Consolidates and corrects several ABI-layer wrappers:

- math_abi/unistd_abi: the scalb implementation is extracted out of
  unistd_abi into math_abi::scalb_svid_impl and called from both entry
  points, so there is a single source of truth. Its doc contract is
  upgraded to the full glibc SVID behavior: non-integer exponent ->
  NaN + FE_INVALID + EDOM, finite nonzero overflow/underflow -> ERANGE,
  and infinite exponents via the x*fn / x/(-fn) forms.
- string_abi / wchar_abi: small correctness adjustments to the wrapper
  surface (error/edge-case handling) kept in lockstep with the core
  implementations.
- fenv_abi / glibc_internal_abi / time_abi: minor fixes and rustfmt
  reflow with no behavioral change.

+++ crates/frankenlibc-abi/src/time_abi.rs
@@ -380,7 +380,7 @@ unsafe fn read_tm(tm: *const libc::tm) -> time_core::BrokenDownTime {
             tm_wday: (*tm).tm_wday,
             tm_yday: (*tm).tm_yday,
             tm_isdst: (*tm).tm_isdst,
-            tm_gmtoff: (*tm).tm_gmtoff as i64,
+            tm_gmtoff: (*tm).tm_gmtoff,
             // tm_zone is NOT dereferenced here: read_tm feeds mktime/timegm/
             // asctime too, and a caller may pass an uninitialised tm_zone to
             // those. Only `strftime` (whose contract reads tm_zone for %Z)
